move some tests to conversions_test.go
